[Chore] estatico-boilerplate: Moved dev styles and scripts into /preview
