+++ packages/estatico-boilerplate/src/preview/layouts/module.hbs
@@ -12,14 +12,11 @@
 	<link rel="stylesheet" href="/assets/css/main{{#unless env.dev}}.min{{/unless}}.css">
 	<link rel="stylesheet" href="/preview/assets/css/main{{#unless env.dev}}.min{{/unless}}.css">
 
-	{{#if env.dev}}
-		<link rel="stylesheet" href="/assets/css/dev.css">
-	{{/if}}
-
 	<script src="/assets/js/head{{#unless env.dev}}.min{{/unless}}.js"></script>
 
 	{{#if env.dev}}
-		<script src="/assets/js/dev.js"></script>
+		<link rel="stylesheet" href="/preview/assets/css/dev.css">
+		<script src="/preview/assets/js/dev.js"></script>
 	{{/if}}
 </head>
 <body {{#if props.svgSprites}} data-svgsprites-options='{{props.svgSprites}}'{{/if}}>
